Koren karta dodata
SquareRootCard je koren karta

+++ Assets/Skripte/AdditionCard.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 public class AdditionCard : AbstractCard {
 
-    public float value = 1;
+    public double value = 1;
 
     public override void Apply (GameManager manager, Player player) {
         player.Lifepoints += value;
     }
 
-    public void SetValue (int newValue) {
+    public void SetValue (double newValue) {
         value = newValue;
 
         Description = $"+{value}";
