+++ Assets/Skripte/DeljenjeCard.cs
@@ -5,14 +5,14 @@ using UnityEngine;
 public class DeljenjeCrad : AbstractCard
 {
 
-    public float value = 1;
+    public double value = 1;
 
     public override void Apply(GameManager manager, Player player)
     {
         player.Lifepoints /= value;
     }
 
-    public void SetValue(int newValue)
+    public void SetValue(double newValue)
     {
         value = newValue;
